Change everything to double and fix 3D GC bugs

+++ src/cage3d.h
@@ -34,8 +34,10 @@ public:
     void buildVertexList(vector<Vector3f> objectVertices);
     void updatePosition();
 
-    bool rayIntersectsTriangle(const Eigen::Vector3f& P, const Eigen::Vector3f& D, const Face& face);
-    bool isPointOutsideMesh(const Eigen::Vector3f& point, HalfEdgeMesh& mesh);
+//    bool rayIntersectsTriangle(const Eigen::Vector3f& P, const Eigen::Vector3f& D, const Face& face);
+//    bool isPointOutsideMesh(const Eigen::Vector3f& point, HalfEdgeMesh& mesh);
+    bool rayIntersectsTriangle(const Eigen::Vector3d& P, const Eigen::Vector3d& D, const Face& face);
+    bool isPointOutsideMesh(const Eigen::Vector3d& point, HalfEdgeMesh& mesh);
 
     // ================== Students, If You Choose To Modify The Code Below, It's On You
